[*] testscripts: dir_changer & execute_wrapper to util.py
